requirements.txt
streamlit
pandas
sqlalchemy
openpyxl
pdfplumber
python-dotenv